Fix head-dedupe (rm code that removed all SSRd style tags)
This was originally for #react-native-stylesheet, but even
that wasn't working since that style el wasn't getting the
prerendercloud-server-side-render attr.

Additionally, even if it was working for react-native, it would
have broke react-native (react-native relies on the existince of those
style tags for insertAdjacent) - so instead this would have had to been
modified to set innerHTML to empty instead of rm'ing the element.

So in conclusion, this style tag removal code just caused problems for
other JS frameworks that were already capable of re-hydrating pre-rendered style els